chore(ci): fix bench drift, widen lint scope, clear rustdoc warnings (#232)
## Summary

- Bench files in `crates/elevator-core/benches/` silently drifted since
the `Accel`/`Speed`/`Weight` newtypes landed in #153 — nightly-bench has
been red while main CI stayed green because `cargo clippy --all-targets`
at `.github/workflows/ci.yml:44` was missing `--workspace` and was only
linting the default member (`elevator-bevy`). This PR closes that gap.
- Adds `benches/common/mod.rs` with a shared `elevator_cfg()` helper so
a future field-type change lands in one place. Each bench now calls
`common::elevator_cfg(id, starting_stop, max_speed, accel, decel,
capacity)`.
- Widens CI's clippy to `--workspace --all-targets --all-features -D
warnings`, and CI's doc check to the full workspace with `RUSTDOCFLAGS:
-D warnings`.
- Adds `cargo check --workspace --all-features --all-targets` to
`.githooks/pre-commit` so this class of drift fails pre-push.
- Fixes the clippy warnings the widened scope surfaces (doc_markdown,
needless_collect, unnecessary_cast, long-literal separators,
doc_lazy_continuation, `Default::default` in a typed context) — mostly
auto-applied by `cargo clippy --fix`, reviewed, committed.
- Fixes three rustdoc warnings: redundant explicit link targets in
`sim.rs:623,627`, broken `ElevatorSim` intra-doc link in
`elevator-gdext/src/lib.rs:4` (added `pub use sim_node::ElevatorSim;`),
private-item links in `elevator-wasm/src/lib.rs:158,171` (unlinked, kept
as prose references to the internal `dto` shape).

No public API changes.

## Test plan

- [x] `cargo check --workspace --all-features --all-targets` — green
- [x] `cargo clippy --workspace --all-targets --all-features -- -D
warnings` — green, zero warnings
- [x] `cargo test --workspace --all-features` — 782 tests pass, 0
failures
- [x] `RUSTDOCFLAGS='-D warnings' cargo doc --no-deps --workspace
--all-features` — green, zero warnings
- [x] `cargo bench -p elevator-core --no-run` — all 5 bench binaries
build
- [x] `.githooks/pre-commit` runs end-to-end green with the new check
- [ ] After merge: manually trigger `Benchmarks (nightly)` via `gh
workflow run bench-nightly.yml` and confirm green baseline

+++ crates/elevator-core/src/tests/move_count_tests.rs
@@ -319,7 +319,7 @@ fn move_count_counts_reposition_arrivals() {
         "expected at least 2 moves (passing + arrival) during reposition, got {count}",
     );
     assert_eq!(
-        count as u64,
+        count,
         sim.metrics().total_moves(),
         "aggregate must equal sum of per-elevator counts",
     );
